refactor: OutputView 인터페이스 생성
