+++ src/main/java/lotto/view/OutputView.java
@@ -5,27 +5,12 @@ import lotto.model.Result;
 
 import java.util.List;
 
-//TODO: 인터페이스를 사용하기 (현재는 파라미터에 어떤 값을 넣을지 정하지 못하였다)
-public class OutputView {
+public interface OutputView {
+    void printBoughtLotto(List<Lotto> boughtLotto);
 
-    public void printBoughtLotto(List<Lotto> boughtLotto) {
-        System.out.printf("%s개를 구매했습니다.%s", boughtLotto.size(), System.lineSeparator());
-        for (Lotto lotto : boughtLotto) {
-            System.out.println(lotto.toString());
-        }
-    }
+    void printWinningStatistics(Result result);
 
-    public void printWinningStatistics(Result result) {
-        System.out.println("당첨 통계");
-        System.out.println("---");
-        result.print();
-    }
+    void printTotalReturn(Double totalReturn);
 
-    public void printTotalReturn(Double totalReturn) {
-        System.out.printf("총 수익률은 %s%%입니다.", String.format("%.1f", totalReturn * 100));
-    }
-
-    public void printErrorMessage(String errorMessage) {
-        System.out.println(errorMessage);
-    }
+    void printErrorMessage(Exception exception);
 }
